Inline dom4j serializer into zatca.go
Fold the dom4j SignedProperties serializer and its tests into the
profile's zatca.go/zatca_test.go rather than separate files.
